Redirect `which` stdout and stderr to /dev/null
kdialog, notify_send and osd_cat notifiers shell out calling the `which`
utility and redirecting standart output to /dev/null. However, when
`which` can't find a file in your $PATH, the error message is
written to standard error, causing it to be displayed when you call the
`notify` method.

This commit makes `which` redirect both its standard output and standard
error to /dev/null.